requirements.txt
streamlit 
pytorch
numpy
albumentations
PIL